Upload the release artefact to a "latest_release" release
We would like to have a stable URL for the most recently released
script, so that we can mention it in the README and people can use it
in update scripts.

Since GitHub's releases/latest URL redirects, and anyway doesn't work
for downloads, we have to create a release called "latest release" and
update it.