Make an image which includes the weave script and its deps
I renamed the weave script to `driver` to distinguish it from the
client script, which simply calls the container with the appropriate
arguments.

The driver script has some modifications so that it will work with
busybox versions of sed. Otherwise, the main change is that it calls
binaries in its filesystem rather than invoking `docker run` for
everthing.

There's a bunch of bind-mounting so that the image has access to
docker, which makes it depend on that being in a certain place
etc. Possibly this could be avoided by separating the docker wrangling
into the client script -- but it may be too entangled.