Add interceptors for create and start
The pre-create hook is a no-op; the post-start hook tells weave to
attach an IP to the container if it's been given a WEAVE_CIDR env
entry.

The proxy is included in the weavetools, and run as `weave
launch-proxy`. One subtlety is that the proc filesystem that is
mounted in the proxy image should be that of the host, not that given
the script. I've hard-wired it to /proc, but it would probably need to
be given as another argument or env entry; or perhaps better, it could
be chosen depending on whether it's run locally or remotely.