weave launch-proxy for using with e.g., swarm
This adds the command `weave launch-proxy`, which makes sure the proxy
is running, and adds the removal of weaveproxy to `weave reset`.

Note on the handling of `/proc`: the environment variable `PROCFS` was
originally added so that the script could be run in a container. The
host's /proc is bind-mounted as /hostproc, and the script must be told
this.

The remote weaving changes got this a bit wrong, by using the value of
`$PROCFS` in a "volume" mount argument (`-v`) to docker -- but when
invoking a container, the left-hand-side of the volume mount should
*always* be `/proc` (and the right-hand-side always `/hostproc`),
since it will always be binding the host's `/proc` to the container's
`/hostproc`.

The launch-proxy command also accept extra arguments to hand to proxy
main.