Account for docker client keep-alive behaviour
When running a container from an image that isn't present on the
server, the docker client will

 1. Ask to create the container and be told the image doesn't exist
 2. Ask to create the image and be fed chunked updates as it's
 downloaded
 3. Ask again to create the container

It likes to keep the connection alive throughout. This causes problems
for us, since as soon as we see the chunked response in 2), we go into
"just forward bytes" mode, and thereby miss intercepting the request
that follows.

To see when the chunked response is finished (and thereby when we
should intercept another request), we need to forward chunks
explicitly. However, once we hijack the connection to go into this
mode, we can't go back to the simplicity of ServeHTTP again since the
server expects us to handle the connection.

To get back to simplicity, we tell the client that we'll be closing
the connection, and close it after the chunked response, forcing it to
open another connection, which we can again intercept, for the next
step.