Reform weave start and smoke test weave {run,start}
`weave start` used to take a CIDR and a containerID, and start the
container then attach to the weave network with the CIDR.

This was simply a convenience, equivalent to `docker start ...; weave
attach ...`. Unlike `docker start`, it did not support -i (hook stdin
up) or -a (hook stdout and stderr up), which we can now do using the
proxy.

It also required the CIDR to be supplied again, whereas we can assume
that the CIDR is already in an environment variable. If it's not,
`weave attach ...` can still be used to the same effect.