Use weave IPAM to give endpoints IPs
This requires the addition of an alloc command to the weave script,
which bypasses checking if the container exists (because we're
supplying an endpoint ID -- the allocator itself doesn't care what you
give it).

I could have called the allocator directly from the plugin, but I'd
need to know the port and IP it's running on, which the script looks
up. (Well, the latter; the former is hard-coded, which poses a
difficulty for running multiple weave routers with IP allocation.)