Switch to libnetwork-compliant implementation
Instead of handing a bridge, interface-to-be-created, and IP to
libcontainer for it to do the work, libnetwork wants drivers to do
most of the work themselves and just hand the interface and IP back.

We also have different structs to serialise. Those in the driver SPI
have net.IP and net.IPNet, which are not serialissable; so, we have to
transform into a string representation (and the driver implementation
has to transform them back).